Se hizo lo que se pudo
